@@ -1,9 +1,16 @@
 {
   "name": "client",
-  "version": "0.13.0",
+  "version": "0.14.0",
   "private": true,
   "devDependencies": {
-    "enzyme": "^2.6.0",
+    "babel-eslint": "^8.0.0",
+    "enzyme": "^2.9.1",
+    "eslint": "^4.6.1",
+    "eslint-config-airbnb": "^15.1.0",
+    "eslint-plugin-import": "^2.7.0",
+    "eslint-plugin-jest": "^21.0.2",
+    "eslint-plugin-jsx-a11y": "^5.1.1",
+    "eslint-plugin-react": "^7.3.0",
     "grunt": "^1.0.1",
     "grunt-available-tasks": "^0.6.3",
     "grunt-string-replace": "^1.3.1",
@@ -12,10 +19,11 @@
     "redux-test-utils": "^0.1.2"
   },
   "dependencies": {
+    "classnames": "^2.2.5",
     "isemail": "^3.0.0",
     "prop-types": "^15.5.10",
-    "react": "^15.4.1",
-    "react-dom": "^15.4.1",
+    "react": "^15.6.1",
+    "react-dom": "^15.6.1",
     "react-redux": "^5.0.6",
     "react-router-dom": "^4.2.2",
     "redux": "^3.7.2"
@@ -24,7 +32,8 @@
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test --env=jsdom",
-    "eject": "react-scripts eject"
+    "eject": "react-scripts eject",
+    "lint": "eslint src"
   },
   "proxy": "http://localhost:8888/"
 }
