@@ -12,7 +12,6 @@ export const SET_LOGIN_STATUS = 'SET_LOGIN_STATUS';
  * Action Creators
  *
  */
-
 export function setToken(token) {
   return {
     type: SET_TOKEN,
@@ -21,6 +20,7 @@ export function setToken(token) {
 }
 
 // TODO: comment
+// TODO: perhaps both of these actions can be combined in the future...
 export function setLoginStatus(isLoggedOut) {
   return {
     type: SET_LOGIN_STATUS,
