@@ -7,17 +7,13 @@ import './Dashboard.css';
 /**
  * @description Dashboard component represents the application dashboard
  *
- * @param {object} props - Component props
- *
  * @returns {<Dashboard />}
  */
-const Dashboard = (props) => {
-  return (
-    <div className="col-sm-9 col-sm-offset-3 col-md-10 col-md-offset-2 main">
-      <h1 className="page-header">Dashboard</h1>
-      <p>There will be things here...</p>
-    </div>
-  );
-}
+const Dashboard = () => (
+  <div className="col-sm-9 col-sm-offset-3 col-md-10 col-md-offset-2 main">
+    <h1 className="page-header">Dashboard</h1>
+    <p>There will be things here...</p>
+  </div>
+);
 
 export default Dashboard;
